requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0
django-ckeditor==6.2.0
django-cors-headers==3.10.1
django-countries==7.3.2
django-heroku==0.0.0
django-js-asset==1.2.2
django-storages==1.12.3
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
email-validator==1.1.3
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
psycopg2-binary==2.9.9
PyJWT==2.3.0
python-dotenv==1.0.1
pytz==2021.3
requests==2.26.0
sentry-sdk==1.5.1
sqlparse==0.4.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
